requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
diffusers==0.34.0
transformers==4.55.2
huggingface-hub==0.34.4
safetensors==0.6.2
tokenizers==0.21.4
sentencepiece==0.2.1
numpy==1.26.4
pillow==11.3.0
tqdm==4.67.1
requests==2.32.4
python-dotenv==1.1.1
